+++ day_17/src/context.cr
@@ -1,12 +1,12 @@
 class Context
   getter registers : Hash(Char, Register)
-  getter output : Array(Int32)
-  property instruction_pointer : Int32
+  getter output : Array(Int128)
+  property instruction_pointer : Int128
 
   def initialize(
     @registers = {} of Char => Register,
     @instruction_pointer = 0,
-    @output = [] of Int32)
+    @output = [] of Int128)
   end
 end
