+++ day_17/src/literal_operand.cr
@@ -1,5 +1,5 @@
 class LiteralOperand < Operand
-  def value : Int32
+  def value : Int128
     @raw_operand
   end
 end
